python/felderize/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "felderize"
version = "0.1.0"
description = "Spark SQL to Feldera SQL translator agent"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.39.0",
    "openai>=1.50.0",
    "httpx>=0.27.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools.package-dir]
felderize = "spark"

[tool.setuptools.package-data]
felderize = [
    "data/skills/**/*.md",
    "data/samples/*.md",
    "data/demo/*.sql",
    "data/demo/expected/*.sql",
    "data/docs/sql/*.md",
    "data/compiler/sql-to-dbsp",
    "data/compiler/*.jar",
]

[project.scripts]
felderize = "felderize.cli:cli"
